Only match on windows with normal/dialog type
When I created a `jumpapp nemo` shortcut, I noticed it was taking me to
the desktop window, instead of launching a new file explorer. This patch
examines the _NET_WM_WINDOW_TYPE to filter out desktop windows and other
special windows. Notice that in the case of nemo, this means you want to
run it with `jumpapp -f nemo`, because the desktop is always running.