Refactor head logic; test through main function
The previous tests were well written, however they were testing helper functions directly and the integration between that function and the rest of the script didn't appear to be tested. I wasn't initially sure if I was going to break something by moving where the `head` call was being made, so I re-wrote the tests to give me more confidence in the change.