Clarify documented app behvior
As pointed out in: https://github.com/mkropat/jumpapp/issues/68

The previous wording was misleading/incorrect. Rather than try to convey
the concept of stacking order in the short synopsis, I intentionally
over-simplify the behavior to a first-approximation explanation in order
to keep things short and undertandable for the first-time reader. I then
updated the introduction to start to hint at the full stacking order
behavior. In general, I think the stacking order behavior is fairly
intuitive, so I'm not too worried about documenting it exactly so long
as it's not actively misleading.